requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.37.0
pydantic==2.11.9
python-multipart==0.0.20
pdfplumber==0.11.4
rapidfuzz==3.9.7
openai==1.51.2
python-docx
pypdf
httpx==0.27.2
httpcore==1.0.5
jinja2==3.1.4